deps: fix requirements.txt
My local packages were out of sync with what was in the requirements.txt. I realized after deleting the local 'env' folder. Added 'request-mock' and pinned Werkzeug to 2.0.3 based on recommnedation from [this StackOverflow article](https://stackoverflow.com/questions/71661851/typeerror-init-got-an-unexpected-keyword-argument-as-tuple) because of an issue I was having with 'as_tuple' when I was running the tests locally.