fix: improve forvo scraping
I started getting 403 errors using urllib.request for the first forvo request, so I updated it to use requests.get, and it seems to be working. I also noticed that when running locally with Docker, some of the animations weren't working when the port was mapped to 80. I set it back to 3000 for now. I only changed it to 80 to make it easier when hosting on the EC2. When the port binding was to 3000, I had to update the security groups to allow traffic on 3000.