fix: update word reference scraping native example sentences
I think Word Reference had an update where instead of using an 'i' tag, they're using a class with a CSS style to apply the italics. This was breaking the nativeExampleSentences. I updated the code, and added a test for vérifier, which was a word that was failing previously.